+++ packages/models/src/index.ts
@@ -16,6 +16,7 @@ export * from "./eservice/eservice.js";
 export * from "./eservice/eserviceReadModelAdapter.js";
 export * from "./eservice/protobufConverterFromV1.js";
 export * from "./eservice/protobufConverterFromV2.js";
+export * from "./eservice/protobufConverterToV2.js";
 export * from "./eservice/eserviceEvents.js";
 
 export * from "./institution/institution.js";
